# Add compression to disk queue