--- a/libbeat/publisher/queue/diskqueue/config.go
+++ b/libbeat/publisher/queue/diskqueue/config.go
@@ -71,6 +71,9 @@ type Settings struct {
 
 	// EncryptionKey is used to encrypt data if SchemaVersion 2 is used.
 	EncryptionKey []byte
+
+	// UseCompression enables or disables LZ4 compression
+	UseCompression bool
 }
 
 // userConfig holds the parameters for a disk queue that are configurable