… NOTICE.txt
… go.sum
… libbeat/publisher/queue/diskqueue/docs/on-disk-structures.md
… libbeat/publisher/queue/diskqueue/enc_compress_test.go