## What does this PR do?

Adds option for compression to beats disk queue.  Works with or without encryption.

## Why is it important?

For certain workloads this can dramatically reduce the amount of data written to disk reducing both encryption overhead and disk I/O

## Checklist

- [x] My code follows the style guidelines of this project
- [x] I have commented my code, particularly in hard-to-understand areas
- [x] I have made corresponding changes to the documentation
~~- [ ] I have made corresponding change to the default configuration files~~
- [x] I have added tests that prove my fix is effective or that my feature works
- [ ] I have added an entry in `CHANGELOG.next.asciidoc` or `CHANGELOG-developer.next.asciidoc`.

## How to test this PR locally

```
go test -bench=1M -benchtime 1x -count 10 -timeout 600m -benchmem > results.txt

benchstat results.txt
name                       time/op
Async1M-16                  21.9s ± 1%
EncryptAsync1M-16           20.9s ± 1%
CompressAsync1M-16          12.4s ± 0%
EncryptCompressAsync1M-16   12.4s ± 1%
Sync1M-16                   25.0s ± 1%
EncryptSync1M-16            27.3s ± 0%
CompressSync1M-16           14.2s ± 0%
EncryptCompressSync1M-16    14.2s ± 0%

name                       alloc/op
Async1M-16                 3.43GB ± 0%
EncryptAsync1M-16          3.43GB ± 0%
CompressAsync1M-16         7.51GB ± 0%
EncryptCompressAsync1M-16  7.51GB ± 0%
Sync1M-16                  3.42GB ± 0%
EncryptSync1M-16           3.42GB ± 0%
CompressSync1M-16          3.52GB ± 0%
EncryptCompressSync1M-16   3.53GB ± 0%

name                       allocs/op
Async1M-16                  40.6M ± 0%
EncryptAsync1M-16           40.4M ± 0%
CompressAsync1M-16          36.9M ± 0%
EncryptCompressAsync1M-16   36.9M ± 0%
Sync1M-16                   39.5M ± 0%
EncryptSync1M-16            40.1M ± 0%
CompressSync1M-16           36.9M ± 0%
EncryptCompressSync1M-16    36.9M ± 0%
```

Also comparing against a run done before this PR, show no significant change in runtime or allocations.

```
benchstat results_previous.txt results.txt

name               old time/op    new time/op    delta
Async1M-16            22.0s ± 1%     21.9s ± 1%    ~     (p=0.436 n=10+10)
EncryptAsync1M-16     20.5s ± 1%     20.9s ± 1%  +1.93%  (p=0.000 n=10+10)
Sync1M-16             25.2s ± 1%     25.0s ± 1%  -0.85%  (p=0.000 n=10+9)
EncryptSync1M-16      27.5s ± 0%     27.3s ± 0%  -0.85%  (p=0.000 n=10+8)

name               old alloc/op   new alloc/op   delta
Async1M-16           3.43GB ± 0%    3.43GB ± 0%    ~     (p=0.529 n=10+10)
EncryptAsync1M-16    3.43GB ± 0%    3.43GB ± 0%    ~     (p=0.089 n=10+10)
Sync1M-16            3.42GB ± 0%    3.42GB ± 0%  +0.00%  (p=0.004 n=10+9)
EncryptSync1M-16     3.42GB ± 0%    3.42GB ± 0%    ~     (p=0.666 n=9+9)

name               old allocs/op  new allocs/op  delta
Async1M-16            40.6M ± 0%     40.6M ± 0%    ~     (p=0.684 n=10+10)
EncryptAsync1M-16     40.5M ± 0%     40.4M ± 0%  -0.19%  (p=0.000 n=10+10)
Sync1M-16             39.5M ± 0%     39.5M ± 0%  +0.05%  (p=0.010 n=10+9)
EncryptSync1M-16      40.1M ± 0%     40.1M ± 0%    ~     (p=0.931 n=9+9)

```

## Related issues

- Relates #31961 
- Relates elastic/elastic-agent-shipper#33